requirements.txt
APScheduler==3.6.3
CacheControl==0.12.6
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
chromedriver-binary==80.0.3987.106.0
Click==7.0
Faker==4.0.1
fastapi==0.52.0
firebase-admin==4.0.0
google-api-core==1.16.0
google-api-python-client==1.7.11
google-auth==1.11.2
google-auth-httplib2==0.0.3
google-cloud-core==1.3.0
google-cloud-firestore==1.6.2
google-cloud-storage==1.26.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
grpcio==1.27.2
h11==0.9.0
httplib2==0.17.0
httptools==0.1.1
idna==2.9
msgpack==1.0.0
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.4
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
rsa==4.0
selenium==3.141.0
six==1.14.0
starlette==0.13.2
text-unidecode==1.3
tzlocal==2.0.0
uritemplate==3.0.1
urllib3==1.25.8
uvicorn==0.11.3
uvloop==0.14.0
websockets==8.1
